Refactor: Extract Service logic from broker to services.py
