+++ tests/tests_new/test_fan_handler.py
@@ -234,11 +234,12 @@ async def test_update_fan_params_sequence(
     tiny_schema = ["11", "22"]
 
     with (
-        patch("custom_components.ramses_cc.broker._2411_PARAMS_SCHEMA", tiny_schema),
+        patch("custom_components.ramses_cc.services._2411_PARAMS_SCHEMA", tiny_schema),
         patch("asyncio.sleep", new_callable=AsyncMock),
     ):
         call_data = {"device_id": FAN_ID}
-        await mock_broker._async_run_fan_param_sequence(call_data)
+        # UPDATED: Call method on service_handler, not broker
+        await mock_broker.service_handler._async_run_fan_param_sequence(call_data)
 
     assert mock_gateway.async_send_cmd.call_count == 2
 
@@ -368,11 +369,13 @@ async def test_run_fan_param_sequence_bad_data(
 
     bad_data_obj = NotADict()
 
+    # UPDATED: Patch on services.RamsesServiceHandler, not broker
     with patch(
-        "custom_components.ramses_cc.broker.RamsesBroker._normalize_service_call",
+        "custom_components.ramses_cc.services.RamsesServiceHandler._normalize_service_call",
         return_value=bad_data_obj,
     ):
-        await mock_broker._async_run_fan_param_sequence({})
+        # UPDATED: Call via service_handler
+        await mock_broker.service_handler._async_run_fan_param_sequence({})
     pass
 
     bad_data_mock = MagicMock(spec=dict)
@@ -380,10 +383,11 @@ async def test_run_fan_param_sequence_bad_data(
     bad_data_mock.get.return_value = "0A"
 
     with patch(
-        "custom_components.ramses_cc.broker.RamsesBroker._normalize_service_call",
+        "custom_components.ramses_cc.services.RamsesServiceHandler._normalize_service_call",
         return_value=bad_data_mock,
     ):
-        await mock_broker._async_run_fan_param_sequence({})
+        # UPDATED: Call via service_handler
+        await mock_broker.service_handler._async_run_fan_param_sequence({})
 
     assert mock_gateway.async_send_cmd.call_count >= 0
 
@@ -562,8 +566,8 @@ async def test_target_to_device_id_single_string(
     # Target dict with a single string device_id (triggers line 1010)
     target = {"device_id": ha_device_id}
 
-    # Patch the device registry getter used in broker.py
-    with patch("custom_components.ramses_cc.broker.dr.async_get") as mock_dr_get:
+    # Patch the device registry getter used in services.py
+    with patch("custom_components.ramses_cc.services.dr.async_get") as mock_dr_get:
         mock_reg = mock_dr_get.return_value
 
         # Create a mock device entry that links HA ID to Ramses ID
@@ -575,28 +579,24 @@ async def test_target_to_device_id_single_string(
             lambda x: mock_entry if x == ha_device_id else None
         )
 
-        # Execute the method
-        result = mock_broker._target_to_device_id(target)
+        # Execute the method on service_handler
+        result = mock_broker.service_handler._target_to_device_id(target)
 
-    # Verify the resolution worked (meaning the string was successfully converted to a list and processed)
+    # Verify the resolution worked
     assert result == ramses_dev_id
 
 
 async def test_target_to_device_id_single_area_string(
     mock_broker: RamsesBroker,
 ) -> None:
-    """Test _target_to_device_id when area_id is a single string.
-
-    This covers the case where a single area_id string is provided in the target,
-    triggering the list conversion logic (lines 1017-1018).
-    """
+    """Test _target_to_device_id when area_id is a single string."""
     area_id = "living_room"
     ramses_dev_id = "10:654321"
 
     target = {"area_id": area_id}
 
     # Patch device registry
-    with patch("custom_components.ramses_cc.broker.dr.async_get") as mock_dr_get:
+    with patch("custom_components.ramses_cc.services.dr.async_get") as mock_dr_get:
         mock_reg = mock_dr_get.return_value
 
         # Create a mock device entry in the correct area with a RAMSES ID
@@ -607,16 +607,16 @@ async def test_target_to_device_id_single_area_string(
         # dev_reg.devices.values() is iterated
         mock_reg.devices.values.return_value = [mock_entry]
 
-        # Execute
-        result = mock_broker._target_to_device_id(target)
+        # Execute on service_handler
+        result = mock_broker.service_handler._target_to_device_id(target)
 
     assert result == ramses_dev_id
 
 
 async def test_target_empty(mock_broker: RamsesBroker) -> None:
     """Test _target_to_device_id with empty or None input."""
-    assert mock_broker._target_to_device_id({}) is None
-    assert mock_broker._target_to_device_id(None) is None
+    assert mock_broker.service_handler._target_to_device_id({}) is None
+    assert mock_broker.service_handler._target_to_device_id(None) is None
 
 
 async def test_target_entity_id_resolution(mock_broker: RamsesBroker) -> None:
@@ -628,8 +628,8 @@ async def test_target_entity_id_resolution(mock_broker: RamsesBroker) -> None:
     ha_dev_id = "ha_dev_1"
 
     with (
-        patch("custom_components.ramses_cc.broker.er.async_get") as mock_er_get,
-        patch("custom_components.ramses_cc.broker.dr.async_get") as mock_dr_get,
+        patch("custom_components.ramses_cc.services.er.async_get") as mock_er_get,
+        patch("custom_components.ramses_cc.services.dr.async_get") as mock_dr_get,
     ):
         # Setup Entity Registry Mock
         mock_ent_reg = mock_er_get.return_value  # This is the registry
@@ -644,10 +644,14 @@ async def test_target_entity_id_resolution(mock_broker: RamsesBroker) -> None:
         mock_dev_reg.async_get.return_value = mock_dev_entry  # Registry returns entry
 
         # Test Single String
-        assert mock_broker._target_to_device_id(target_single) == ramses_id
+        assert (
+            mock_broker.service_handler._target_to_device_id(target_single) == ramses_id
+        )
 
         # Test List
-        assert mock_broker._target_to_device_id(target_list) == ramses_id
+        assert (
+            mock_broker.service_handler._target_to_device_id(target_list) == ramses_id
+        )
 
 
 async def test_target_device_id_resolution(mock_broker: RamsesBroker) -> None:
@@ -657,7 +661,7 @@ async def test_target_device_id_resolution(mock_broker: RamsesBroker) -> None:
 
     ramses_id = "02:222222"
 
-    with patch("custom_components.ramses_cc.broker.dr.async_get") as mock_dr_get:
+    with patch("custom_components.ramses_cc.services.dr.async_get") as mock_dr_get:
         # Setup Device Registry Mock
         mock_dev_reg = mock_dr_get.return_value
         mock_dev_entry = MagicMock()
@@ -665,10 +669,14 @@ async def test_target_device_id_resolution(mock_broker: RamsesBroker) -> None:
         mock_dev_reg.async_get.return_value = mock_dev_entry
 
         # Test Single String
-        assert mock_broker._target_to_device_id(target_single) == ramses_id
+        assert (
+            mock_broker.service_handler._target_to_device_id(target_single) == ramses_id
+        )
 
         # Test List
-        assert mock_broker._target_to_device_id(target_list) == ramses_id
+        assert (
+            mock_broker.service_handler._target_to_device_id(target_list) == ramses_id
+        )
 
 
 async def test_target_priority_order(mock_broker: RamsesBroker) -> None:
@@ -682,8 +690,8 @@ async def test_target_priority_order(mock_broker: RamsesBroker) -> None:
     id_from_entity = "01:000001"
 
     with (
-        patch("custom_components.ramses_cc.broker.er.async_get") as mock_er_get,
-        patch("custom_components.ramses_cc.broker.dr.async_get") as mock_dr_get,
+        patch("custom_components.ramses_cc.services.er.async_get") as mock_er_get,
+        patch("custom_components.ramses_cc.services.dr.async_get") as mock_dr_get,
     ):
         # 1. Setup successful Entity Lookup
         mock_ent_reg = mock_er_get.return_value
@@ -704,14 +712,16 @@ async def test_target_priority_order(mock_broker: RamsesBroker) -> None:
         mock_dev_reg.async_get.side_effect = side_effect
 
         # Should return the one found via entity_id, ignoring device_id/area_id logic
-        assert mock_broker._target_to_device_id(target) == id_from_entity
+        assert (
+            mock_broker.service_handler._target_to_device_id(target) == id_from_entity
+        )
 
 
 async def test_target_resolution_failures(mock_broker: RamsesBroker) -> None:
     """Test that it returns None when lookups fail or entries lack correct domain."""
     target = {"device_id": "ha_dev_bad"}
 
-    with patch("custom_components.ramses_cc.broker.dr.async_get") as mock_dr_get:
+    with patch("custom_components.ramses_cc.services.dr.async_get") as mock_dr_get:
         mock_dev_reg = mock_dr_get.return_value
 
         # Case 1: Device entry found, but no RAMSES domain identifier
@@ -719,18 +729,23 @@ async def test_target_resolution_failures(mock_broker: RamsesBroker) -> None:
         mock_entry.identifiers = {("other_domain", "some_id")}
         mock_dev_reg.async_get.return_value = mock_entry
 
-        assert mock_broker._target_to_device_id(target) is None
+        assert mock_broker.service_handler._target_to_device_id(target) is None
 
         # Case 2: Device entry is None (device not found in registry)
         mock_dev_reg.async_get.return_value = None
-        assert mock_broker._target_to_device_id(target) is None
+        assert mock_broker.service_handler._target_to_device_id(target) is None
 
     # Case 3: Entity found, but has no device_id
-    with patch("custom_components.ramses_cc.broker.er.async_get") as mock_er_get:
+    with patch("custom_components.ramses_cc.services.er.async_get") as mock_er_get:
         mock_ent_reg = mock_er_get.return_value
         mock_ent_reg.async_get.return_value = MagicMock(device_id=None)
 
-        assert mock_broker._target_to_device_id({"entity_id": "sensor.orphan"}) is None
+        assert (
+            mock_broker.service_handler._target_to_device_id(
+                {"entity_id": "sensor.orphan"}
+            )
+            is None
+        )
 
 
 async def test_get_fan_param_fallback_hgi(
